[target-truffle-v4] Refactor codegen
